requirements.txt
# PDF-based RAG Chatbot Dependencies

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# PDF Processing
pdfplumber>=0.10.0

# Embeddings & ML
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Store
faiss-cpu>=1.7.4

# LLM API
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
